+++ VoidEngine/VoidEngine/Camera.hpp
@@ -20,7 +20,7 @@ public:
     void            setPosition(glm::vec3);
     void            setTarget(glm::vec3);
     void            setUp(glm::vec3);
-	virtual void	handleMouseMovement(int x, int y);
+	virtual void	handleMouseMovement(float x, float y);
 	virtual void	handleKeyboard(char input, float deltaTime);
 
     // Projection Functions
